refactor: inverse dependency for `postTrack` function => move provisional `postTrack` implem to dedicated file
