+++ app/lib/my-http-wrapper/http/Application.js
@@ -6,6 +6,9 @@ const express = require('express');
 const formidable = require('formidable');
 const qset = require('q-set'); // instead of body-parser, for form fields with brackets
 const sessionTracker = require('../../../controllers/admin/session.js');
+const {
+  postTrack,
+} = require('../../../api-v2/provisional-features/postTrack.js');
 
 const LOG_THRESHOLD = parseInt(process.env.LOG_REQ_THRESHOLD_MS ?? '1000', 10);
 
@@ -141,10 +144,14 @@ exports.Application = class Application {
 
     if (this._features.auth) {
       // Openwhyd API V2 relies on OAuth/Auth0
-      require('../../../api-v2/OpenwhydApiV2.js').injectOpenwhydAPIV2(app, {
-        issuerBaseURL: process.env.AUTH0_ISSUER_BASE_URL, // identifier of the Auth0 account
-        urlPrefix: process.env.WHYD_URL_PREFIX, // identifier of Openwhyd API v2, as set on Auth0
-      });
+      require('../../../api-v2/OpenwhydApiV2.js').injectOpenwhydAPIV2(
+        app,
+        {
+          issuerBaseURL: process.env.AUTH0_ISSUER_BASE_URL, // identifier of the Auth0 account
+          urlPrefix: process.env.WHYD_URL_PREFIX, // identifier of Openwhyd API v2, as set on Auth0
+        },
+        { postTrack },
+      );
     }
 
     attachLegacyRoutesFromFile(
